+++ public/sw.js
@@ -1,3 +1,4 @@
+import { pageCache } from 'workbox-recipes';
 import { CacheableResponsePlugin } from 'workbox-cacheable-response';
 import { ExpirationPlugin } from 'workbox-expiration';
 import * as navigationPreload from 'workbox-navigation-preload';
@@ -11,6 +12,11 @@ import {
 navigationPreload.enable();
 
 self.__WB_DISABLE_DEV_LOGS = true;
+
+// Cache HTML pages
+pageCache({
+  warmCache: ['./compose/'],
+});
 
 // Custom plugin to manage hashed assets
 class AssetHashPlugin {
@@ -167,9 +173,8 @@ const assetsRoute = new Route(
     const hasHash = /-[0-9a-z-]{4,}\./i.test(request.url);
     return sameOrigin && isAsset && hasHash;
   },
-  new NetworkFirst({
+  new StaleWhileRevalidate({
     cacheName: 'assets',
-    networkTimeoutSeconds: 5,
     plugins: [
       // Only enable AssetHashPlugin in production
       ...(import.meta.env.PROD
@@ -235,6 +240,28 @@ const apiExtendedRoute = new RegExpRoute(
   }),
 );
 registerRoute(apiExtendedRoute);
+
+// Cache ActivityPub requests (Accept: application/activity+json)
+const activityPubRoute = new Route(
+  ({ request }) => {
+    const acceptHeader = request.headers.get('accept');
+    return acceptHeader?.includes('application/activity+json');
+  },
+  new StaleWhileRevalidate({
+    cacheName: 'activity-json',
+    plugins: [
+      new ExpirationPlugin({
+        maxEntries: 30,
+        maxAgeSeconds: 60 * 60, // 1 hour
+        ...expirationPluginOptions,
+      }),
+      new CacheableResponsePlugin({
+        statuses: [0, 200],
+      }),
+    ],
+  }),
+);
+registerRoute(activityPubRoute);
 
 // Note: expiration is not working as expected
 // https://github.com/GoogleChrome/workbox/issues/3316