+++ vite.config.js
@@ -28,13 +28,18 @@ const {
 
 const now = new Date();
 let commitHash;
+let commitTime;
 let fakeCommitHash = false;
 try {
-  commitHash = execSync('git rev-parse --short HEAD').toString().trim();
+  const gitResult = execSync('git log -1 --format="%h %cI"').toString().trim();
+  const [hash, time] = gitResult.split(' ');
+  commitHash = hash;
+  commitTime = new Date(time);
 } catch (error) {
   // If error, means git is not installed or not a git repo (could be downloaded instead of git cloned)
   // Fallback to random hash which should be different on every build run 🤞
   commitHash = uid();
+  commitTime = now;
   fakeCommitHash = true;
 }
 
@@ -52,10 +57,16 @@ export default defineConfig({
   define: {
     __BUILD_TIME__: JSON.stringify(now),
     __COMMIT_HASH__: JSON.stringify(commitHash),
+    __COMMIT_TIME__: JSON.stringify(commitTime),
     __FAKE_COMMIT_HASH__: fakeCommitHash,
   },
   server: {
     host: true,
+    watch: {
+      awaitWriteFinish: {
+        pollInterval: 1000,
+      },
+    },
   },
   css: {
     preprocessorMaxWorkers: 1,
@@ -130,6 +141,21 @@ export default defineConfig({
           ]
         : []),
     ]),
+    {
+      // https://developers.cloudflare.com/pages/configuration/early-hints/
+      name: 'generate-headers',
+      writeBundle(_, bundle) {
+        const cssFiles = Object.keys(bundle).filter((file) =>
+          file.endsWith('.css'),
+        );
+        if (cssFiles.length > 0) {
+          const links = cssFiles
+            .map((file) => `  Link: <${file}>; rel=preload; as=style`)
+            .join('\n');
+          fs.writeFileSync(resolve(__dirname, 'dist/_headers'), `/\n${links}`);
+        }
+      },
+    },
     VitePWA({
       manifest: {
         name: CLIENT_NAME,
@@ -173,14 +199,38 @@ export default defineConfig({
       brotli: true,
       open: false,
     }),
+    {
+      name: 'css-ordering-plugin',
+      transformIndexHtml(html) {
+        const stylesheets = [];
+        html = html.replace(
+          /<link[^>]*rel=["']stylesheet["'][^>]*>/g,
+          (match) => {
+            stylesheets.push(match);
+            return '';
+          },
+        );
+
+        // Try to place before first <link> tag, fallback to after last <meta> tag
+        const linkRegex = /<link[^>]*>/;
+        if (linkRegex.test(html)) {
+          return html.replace(linkRegex, (match) => {
+            return stylesheets.join('') + match;
+          });
+        } else {
+          return html.replace(/(<meta[^>]*>)(?![\s\S]*<meta)/, (match) => {
+            return match + stylesheets.join('');
+          });
+        }
+      },
+    },
   ],
   build: {
     sourcemap: true,
-    // Note: In Vite 6, if cssCodeSplit = false, it will show error "Cannot read properties of undefined (reading 'includes')"
-    // TODO: Revisit this when this issue is fixed
-    // cssCodeSplit: false,
+    cssCodeSplit: false,
     rollupOptions: {
       treeshake: false,
+      external: ['@xmldom/xmldom'], // exifreader's optional dependency, not needed
       input: {
         main: resolve(__dirname, 'index.html'),
         compose: resolve(__dirname, 'compose/index.html'),